requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
python-dotenv
httpx
dashscope

